requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest-playwright==0.7.0
pytest-html==4.1.1
pytest-base-url==2.1.0
python-dotenv==1.0.1
